@@ -1,7 +1,8 @@
 module org.monarchinitiative.svart {
     exports org.monarchinitiative.svart;
     exports org.monarchinitiative.svart.assembly;
+    exports org.monarchinitiative.svart.sequence;
     exports org.monarchinitiative.svart.interval;
-    exports org.monarchinitiative.svart.parsers;
-    exports org.monarchinitiative.svart.util;
+    exports org.monarchinitiative.svart.variantkey;
+    exports org.monarchinitiative.svart.vcf;
 }
