@@ -6,7 +6,7 @@ import java.util.*;
 
 
 /**
- * Sorted array of {@link Interval} objects representing an immutable interval
+ * Sorted array of {@link IntervalTreeNode} objects representing an immutable interval
  * tree.
  * <p>
  * The query results are sorted lexicographically by <code>(begin, end)</code>.
@@ -16,40 +16,40 @@ import java.util.*;
  * @author <a href="mailto:manuel.holtgrewe@charite.de">Manuel Holtgrewe</a>
  * @since 2.0.0
  */
-public class IntervalTree<T> {
+public final class IntervalTree<T> {
 
     public static final CoordinateSystem COORDINATE_SYSTEM = CoordinateSystem.ZERO_BASED;
 
     /**
-     * list of {@link Interval} objects, sorted by begin position
+     * list of {@link IntervalTreeNode} objects, sorted by begin position
      */
-    private final List<Interval<T>> intervals;
+    private final List<IntervalTreeNode<T>> nodes;
 
     /**
-     * list of {@link Interval} objects, sorted by end position
+     * list of {@link IntervalTreeNode} objects, sorted by end position
      */
-    private final List<Interval<T>> intervalsEnd;
+    private final List<IntervalTreeNode<T>> intervalsEnd;
 
     /**
      * Construct object with the given values.
      */
     public IntervalTree(Collection<T> elements, IntervalNormaliser<T> intervalNormaliser) {
         IntervalListBuilder<T> pair = new IntervalListBuilder<>(elements, intervalNormaliser);
-        this.intervals = pair.intervals;
+        this.nodes = pair.nodes;
         this.intervalsEnd = pair.intervalsEnd;
     }
 
     /**
-     * @return {@link Interval}s, sorted by begin position
+     * @return {@link IntervalTreeNode}s, sorted by begin position
      */
-    public List<Interval<T>> intervalsByStart() {
-        return intervals;
+    public List<IntervalTreeNode<T>> intervalsByStart() {
+        return nodes;
     }
 
     /**
-     * @return {@link Interval}s, sorted by end position
+     * @return {@link IntervalTreeNode}s, sorted by end position
      */
-    public List<Interval<T>> intervalsByEnd() {
+    public List<IntervalTreeNode<T>> intervalsByEnd() {
         return intervalsEnd;
     }
 
@@ -57,7 +57,7 @@ public class IntervalTree<T> {
      * @return the number of elements in the tree
      */
     public int size() {
-        return intervals.size();
+        return nodes.size();
     }
 
     /**
@@ -70,7 +70,7 @@ public class IntervalTree<T> {
      */
     public IntervalOverlaps<T> findOverlappingWithPoint(int point) {
         List<T> overlapping = new ArrayList<>();
-        findOverlappingWithPoint(0, intervals.size(), intervals.size() / 2, point, overlapping);
+        findOverlappingWithPoint(0, nodes.size(), nodes.size() / 2, point, overlapping);
 
         // if overlapping interval was found then return this set
         if (!overlapping.isEmpty()) {
@@ -87,26 +87,26 @@ public class IntervalTree<T> {
      * @return right neighbor of the given point if any, or <code>null</code>
      */
     private T findRightNeighbor(int point) {
-        final Interval<T> query = new Interval<>(point, point, null, point);
-        int idx = Collections.binarySearch(intervals, query, Comparator.comparingInt(Interval::begin));
+        final IntervalTreeNode<T> query = new IntervalTreeNode<>(point, point, null, point);
+        int idx = Collections.binarySearch(nodes, query, Comparator.comparingInt(IntervalTreeNode::begin));
 
         if (idx >= 0) {
             throw new IllegalStateException("Found element although in right neighbor search!");
         }
         idx = -(idx + 1); // convert to insertion point
 
-        if (idx == intervals.size()) {
+        if (idx == nodes.size()) {
             return null;
         }
-        return intervals.get(idx).value();
+        return nodes.get(idx).value();
     }
 
     /**
      * @return left neighbor of the given point if any, or <code>null</code>
      */
     private T findLeftNeighbor(int point) {
-        final Interval<T> query = new Interval<>(point, point, null, point);
-        int idx = Collections.binarySearch(intervalsEnd, query, Comparator.comparingInt(Interval::end));
+        final IntervalTreeNode<T> query = new IntervalTreeNode<>(point, point, null, point);
+        int idx = Collections.binarySearch(intervalsEnd, query, Comparator.comparingInt(IntervalTreeNode::end));
 
         if (idx >= 0) {
             idx += 1;
@@ -121,7 +121,7 @@ public class IntervalTree<T> {
     }
 
     /**
-     * Implementation of in-order traversal of the encoded tree with pruning using {@link Interval#maxEnd()}.
+     * Implementation of in-order traversal of the encoded tree with pruning using {@link IntervalTreeNode#maxEnd()}.
      *
      * @param begin       begin index of subtree to search through
      * @param end         end index of subtree to search through
@@ -135,7 +135,7 @@ public class IntervalTree<T> {
             return;
         }
 
-        Interval<T> node = intervals.get(center); // shortcut to current node
+        IntervalTreeNode<T> node = nodes.get(center); // shortcut to current node
 
         // point is right of the rightmost point of any interval in this node
         if (node.allLeftOf(point)) {
@@ -173,7 +173,7 @@ public class IntervalTree<T> {
      */
     public IntervalOverlaps<T> findOverlappingWithInterval(int begin, int end) {
         List<T> overlapping = new ArrayList<>();
-        findOverlappingWithInterval(0, intervals.size(), intervals.size() / 2, begin, end, overlapping);
+        findOverlappingWithInterval(0, nodes.size(), nodes.size() / 2, begin, end, overlapping);
 
         // if overlapping interval was found then return this set
         if (!overlapping.isEmpty()) {
@@ -187,7 +187,7 @@ public class IntervalTree<T> {
     }
 
     /**
-     * Implementation of in-order traversal of the encoded tree with pruning using {@link Interval#maxEnd()}.
+     * Implementation of in-order traversal of the encoded tree with pruning using {@link IntervalTreeNode#maxEnd()}.
      *
      * @param begin       begin index of subtree to search through
      * @param end         end index of subtree to search through
@@ -203,7 +203,7 @@ public class IntervalTree<T> {
         }
 
         // shortcut to current node
-        Interval<T> node = intervals.get(center);
+        IntervalTreeNode<T> node = nodes.get(center);
 
         // iBegin is right of the rightmost point of any interval in this node
         if (node.allLeftOf(iBegin)) {
@@ -236,12 +236,12 @@ public class IntervalTree<T> {
      */
     private static class IntervalListBuilder<T> {
 
-        private final List<Interval<T>> intervals;
-        private final List<Interval<T>> intervalsEnd;
+        private final List<IntervalTreeNode<T>> nodes;
+        private final List<IntervalTreeNode<T>> intervalsEnd;
 
         public IntervalListBuilder(Collection<T> elements, IntervalNormaliser<T> intervalNormaliser) {
             List<MutableInterval<T>> mutableIntervals = buildMutableIntervals(elements, intervalNormaliser);
-            this.intervals = toIntervalList(mutableIntervals);
+            this.nodes = toIntervalList(mutableIntervals);
             this.intervalsEnd = buildIntervalsEnd(mutableIntervals);
         }
 
@@ -278,18 +278,18 @@ public class IntervalTree<T> {
         /**
          * Fill {@link #intervalsEnd}.
          */
-        private List<Interval<T>> buildIntervalsEnd(List<MutableInterval<T>> tmpList) {
+        private List<IntervalTreeNode<T>> buildIntervalsEnd(List<MutableInterval<T>> tmpList) {
             // sort by (end, begin)
             tmpList.sort(this::compareEndBegin);
             return toIntervalList(tmpList);
         }
 
-        private List<Interval<T>> toIntervalList(List<MutableInterval<T>> tmpList) {
-            List<Interval<T>> intervalList = new ArrayList<>(tmpList.size());
+        private List<IntervalTreeNode<T>> toIntervalList(List<MutableInterval<T>> tmpList) {
+            List<IntervalTreeNode<T>> intervalTreeNodes = new ArrayList<>(tmpList.size());
             for (MutableInterval<T> i : tmpList) {
-                intervalList.add(new Interval<>(i.getBegin(), i.getEnd(), i.getValue(), i.getMaxEnd()));
+                intervalTreeNodes.add(new IntervalTreeNode<>(i.getBegin(), i.getEnd(), i.getValue(), i.getMaxEnd()));
             }
-            return List.copyOf(intervalList);
+            return List.copyOf(intervalTreeNodes);
         }
 
         private int compareEndBegin(MutableInterval<T> o1, MutableInterval<T> o2) {
